requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
psycopg2-binary
joblib
